0.24b: Fixed delta pruning. Optimised variables.

+++ Source/Engine/attack.c
@@ -9,7 +9,7 @@ const int BiDir[4] = { -9, -11, 11, 9 }; // at one square
 const int KiDir[8] = { -1, -10,	1, 10, -9, -11, 11, 9 };
 
 // Returns 1 if a given square is attacked
-int SqAttacked(const int sq, const int side, const S_BOARD *pos) {
+uint8_t SqAttacked(const int sq, const int side, const S_BOARD *pos) {
 
 	int pce,index,t_sq,dir;
