+++ Source/Engine/polybook.c
@@ -29,7 +29,7 @@ void InitPolyBook() {
 		fseek(pFile,0,SEEK_END);
 		long position = ftell(pFile);
 		
-		if(position < sizeof(S_POLY_BOOK_ENTRY)) {
+		if(position < (long)sizeof(S_POLY_BOOK_ENTRY)) {
 			printf("No Entries Found\n");
 			return;
 		}
@@ -42,7 +42,7 @@ void InitPolyBook() {
 		
 		size_t returnValue;
 		returnValue = fread(entries, sizeof(S_POLY_BOOK_ENTRY), NumEntries, pFile);
-		printf("fread() %ld Entries Read in from file\n", returnValue);
+		printf("fread() %llu Entries Read in from file\n", returnValue);
 		
 		if(NumEntries > 0) {
 			EngineOptions->UseBook = TRUE;
